data/requirements.txt
adium-theme-ubuntu==0.3.4
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
ccsm==0.9.12.2
certifi==2017.11.5
chardet==3.0.4
compizconfig-python==0.9.12.2
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==3.1.2
plotly==2.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
requests==2.18.4
scandir==1.6
scikit-learn==0.19.0
scipy==0.19.1
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
stevedore==1.26.0
subprocess32==3.2.7
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
Theano==0.9.0
tornado==4.5.3
traitlets==4.3.2
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.1
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.1.3
